% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_channel_stats_all.R
\name{get_channel_stats_all}
\alias{get_channel_stats_all}
\title{Get statistics of a Channels}
\usage{
get_channel_stats_all(ids, part = c("snippet", "statistics"), ...)
}
\arguments{
\item{ids}{Character. Id of the channel}

\item{part}{What parts should be returned}

\item{\dots}{Additional arguments passed to \code{\link{tuber_GET}}.}
}
\value{
data frame:
\code{kind, etag, id, snippet (list of details of the channel
including title), statistics}

If the \code{ids} is mistyped or there is no information, an empty list is returned
}
\description{
Get statistics of a Channels
}
\examples{
\dontrun{

# Set API token via yt_oauth() first

get_channel_stats_all(ids="UCMtFAi84ehTSYSE9XoHefig")
get_channel_stats(ids="UCMtFAi84ehTSYSE9Xo") # Incorrect channel ID
}
}
\references{
\url{https://developers.google.com/youtube/v3/docs/channels/list}
}
