% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_featured_channels.R
\name{get_featured_channels}
\alias{get_featured_channels}
\title{Get all the featured channels on a channel page}
\usage{
get_featured_channels(channel_id = NULL, ...)
}
\arguments{
\item{channel_id}{string; Required.
\code{channel_id}: channel ID.}

\item{\dots}{Additional arguments passed to \code{\link{tuber_GET}}.}
}
\value{
a \code{data.frame} with the following columns:
\code{kind, etag, id,}
\code{ contentDetails.playlists, contentDetails.channels}
}
\description{
Get all the featured channels on a channel page
}
\examples{
\dontrun{

# Set API token via yt_oauth() first

get_featured_channels(channel_id = "UCG8rbF3g2AMX70yOd8vqIZg")
}
}
\references{
\url{https://developers.google.com/youtube/v3/docs/channelSections}
}
